DetalleMuestreo Modificacion Muestreo

+++ src/main/java/com/siapa/dao/DetalleMuestreoDao.java
@@ -15,6 +15,6 @@ import org.springframework.stereotype.Repository;
  * @author Joao
  */
 @Repository
-public class DetalleMuestreoDao extends GenericDao<DetalleMuestreo, Long> {
+public class DetalleMuestreoDao extends GenericDao<DetalleMuestreo, Integer> {
     
 }
